test(types): move all types to types directory
