+++ tests/dummy/models/PostCommentEmbed.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 
-import type { ModelData } from '../../../src'
+import type { ModelData } from '../../../src/types'
 import BaseModel from './BaseModel'
 import CommentWrapped from './CommentWrapped'
 import TagEmbed from './TagEmbed'
